Fixes #134 + Fixes #139.
Also, fixes an unrecorded bug that caused system stylesets and styles not
to be able to be opened because nothing was added to their "members" when
they were created => when a user registers and system stylesets and styles
are copied, they should get the current user as member. Also, when
performing this copy, the copied stylesets and styles should be marked as
non-system (i.e. user).