Fixes #194 + #195 + #199.
The project (i.e. user) styleset is now COPIED when a document is created, setting the copy as the document's default styleset.

To change a document's default styleset, use `PUT /document:documentId/update`. This will no longer remove the old default styleset from the document's stylesets, because the it makes it easier for the frontend to get all stylesets to be shown in "Typography" in one call (`Document.listStylesets()`)

To add/use/reference a non-default styleset to a document, use `PUT /styleset/:stylesetId/document/:documentId`. This will no longer copy the styleset since this is now done when creating a document.
Thus, this call must ONLY be made with a document stylesets, i.e. NOT with non-copied stylesets (i.e. system stylesets). Adding a user styleset to the document is of course incorrect as a document must only contain document stylesets, i.e. stylesets which are copies of user stylesets. Adding a non-copied styleset this way will make `GET /document/:documentIdPopulatedStylesets/stylesets` return incorrect stylesets, because the user styleset will not be filtered out (because the user styleset's "original" value is undefined). The IT "Applying a(nother) styleset to a document..." starting at line 1276 was removed for this reason because "stylesetId2" was not a copy of a system styleset.

Confer comments in #184.

Also:
* Document.update() does not change (remove from) document.stylesets anymore so no need to test this (it.js, line 777-778, 792-793 and 807-808).
* it.js, line 1304: system styleset should not be returned by `Document.listStylesets()` since it is neither a user or document styleset.