Fixes #234.
Allow frontend to rearrange the combined list of user and document stylesets (in Styleset.update(), it is possible to set the "order" of just one styleset).

We need the "Styleset.order" attribute because document/:id/stylesets/ will return a combination of user and document stylesets, so if we only relied on the frontend sending an array of styleset ids in a specific order (as for documents in a project), document/:id/stylesets/ would not be able to keep this order.

Also:
* Changed script to import system stylesets to sort stylesets alphabetically (process files in series instead of parallel).
* Renamed '/styleset/list' to more appropriate name '/styleset/systemlist'.