Fixes #771.
For EPUB download to function *without* nginx, e.g. for the integration test and running the app without nginx, the "users" and "projects" directories must be located under the "public" folder (this is where express' world starts so the files it is serving must be located under this folder).

This commit makes the frontend (user) download the EPUB from node instead of going via an nginx alias (/home/scripler/prod/data/). The download link in the status email will, (for the production config) continue to use an nginx alias to point to a folder outside "public".

The error currently also in production

    Error: Command failed: ERROR: tmp/app/tmp/e83c2496-0f0e-4bcc-aa52-1949931391f6.epub: Cannot read header
    ERROR: tmp/app/tmp/e83c2496-0f0e-4bcc-aa52-1949931391f6.epub: I/O error: zip file is empty

is *not* fixed by this commit. Issue #812 has been created for this.