Force mongoose version to compatible version.
This is an attempt to fix failing tests in a Strider build giving the "This is an UNSTABLE release of Mongoose" warning, assuming that the tests fail due to this warning.

This fix has already been made in another branch, "issue-403", but this has not yet been merged to the common branch, "beta1".